// src/app/shared/models/skill-tree.ts
import { Skill } from "./skill";

export class SkillTree {
  selectedTierOneSkill: Skill;
  tierOneSkills: Skill[];
  selectedTierTwoSkill: Skill;
  tierTwoSkills: Skill[];
  selectedTierThreeSkill: Skill;
  tierThreeSkills: Skill[];
  selectedTierFourSkill: Skill;
  tierFourSkills: Skill[];
  selectedTierFiveSkill: Skill;
  tierFiveSkills: Skill[];
}